fix: Delete equipment left on planet on expedition failure
Currently, equipment present on the Planet room, most notably the Icarus,
is not deleted on expedition failure (i.e. all members of the expedition
are dead or lost).

This poses an issue where a subsequent (and successful) expedition with a
Patroller retrieves all equipment in the room... including the Icarus.

This fix deletes all equipment left on the planet if the expedition fails
(and the Daedalus does not have the Auto Return Icarus project). This
ensures that the Icarus can no longer be retrieved by further expeditions,
as it no longer exists.

closes #1135